[pdf] FreeText annot fixes

+++ pdf/src/annotations/line.js
@@ -557,7 +557,7 @@
         let nIntent = this.GetIntent();
         if (nIntent != null) {
             memory.annotFlags |= (1 << 20);
-            memory.WriteDouble(nIntent);
+            memory.WriteByte(nIntent);
         }
 
         // leader Line Offset
